Fix score type conversion issues and update player model

+++ core/src/no/ntnu/game/Controllers/GameRoomController.java
@@ -212,12 +212,14 @@ public class GameRoomController {
         }
         if (room.getGameMode().equals(GameRoom.GameMode.FASTEST_KNIGHT)) {
             if (room.getCreatingPlayerState().getScore() < room.getCreatingPlayer().getFastestTime()) {
-                room.getCreatingPlayer().setFastestTime(room.getCreatingPlayerState().getScore());
+                // TODO: Alter score to accept floats
+                room.getCreatingPlayer().setFastestTime((float) room.getCreatingPlayerState().getScore());
                 _FI.savePlayer(room.getCreatingPlayer());
             }
             if (roomType.equals(RoomType.ONLINE)) {
                 if (room.getJoiningPlayerState().getScore() < room.getJoiningPlayer().getFastestTime()) {
-                    room.getJoiningPlayer().setFastestTime(room.getJoiningPlayerState().getScore());
+                    // TODO: Alter score to accept floats
+                    room.getJoiningPlayer().setFastestTime((float) room.getJoiningPlayerState().getScore());
                     _FI.savePlayer(room.getJoiningPlayer());
                 }
             }
